// project-management-frontend/src/@api/models/user-analytics-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Project Management Tool API
 * API documentation for Project Management Tool
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface UserAnalyticsDto
 */
export interface UserAnalyticsDto {
    /**
     * 
     * @type {number}
     * @memberof UserAnalyticsDto
     */
    assignedTasks?: number;
    /**
     * 
     * @type {number}
     * @memberof UserAnalyticsDto
     */
    completedTasks?: number;
    /**
     * 
     * @type {number}
     * @memberof UserAnalyticsDto
     */
    overdueTasks?: number;
    /**
     * 
     * @type {{ [key: string]: number; }}
     * @memberof UserAnalyticsDto
     */
    userTasksByStatus?: { [key: string]: number; };
    /**
     * 
     * @type {number}
     * @memberof UserAnalyticsDto
     */
    totalProjects?: number;
    /**
     * 
     * @type {number}
     * @memberof UserAnalyticsDto
     */
    activeProjects?: number;
    /**
     * 
     * @type {number}
     * @memberof UserAnalyticsDto
     */
    completedProjects?: number;
    /**
     * 
     * @type {number}
     * @memberof UserAnalyticsDto
     */
    totalTeams?: number;
}

/**
 * Check if a given object implements the UserAnalyticsDto interface.
 */
export function instanceOfUserAnalyticsDto(value: object): value is UserAnalyticsDto {
    return true;
}

export function UserAnalyticsDtoFromJSON(json: any): UserAnalyticsDto {
    return UserAnalyticsDtoFromJSONTyped(json, false);
}

export function UserAnalyticsDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): UserAnalyticsDto {
    if (json == null) {
        return json;
    }
    return {
        
        'assignedTasks': json['assignedTasks'] == null ? undefined : json['assignedTasks'],
        'completedTasks': json['completedTasks'] == null ? undefined : json['completedTasks'],
        'overdueTasks': json['overdueTasks'] == null ? undefined : json['overdueTasks'],
        'userTasksByStatus': json['userTasksByStatus'] == null ? undefined : json['userTasksByStatus'],
        'totalProjects': json['totalProjects'] == null ? undefined : json['totalProjects'],
        'activeProjects': json['activeProjects'] == null ? undefined : json['activeProjects'],
        'completedProjects': json['completedProjects'] == null ? undefined : json['completedProjects'],
        'totalTeams': json['totalTeams'] == null ? undefined : json['totalTeams'],
    };
}

export function UserAnalyticsDtoToJSON(json: any): UserAnalyticsDto {
    return UserAnalyticsDtoToJSONTyped(json, false);
}

export function UserAnalyticsDtoToJSONTyped(value?: UserAnalyticsDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'assignedTasks': value['assignedTasks'],
        'completedTasks': value['completedTasks'],
        'overdueTasks': value['overdueTasks'],
        'userTasksByStatus': value['userTasksByStatus'],
        'totalProjects': value['totalProjects'],
        'activeProjects': value['activeProjects'],
        'completedProjects': value['completedProjects'],
        'totalTeams': value['totalTeams'],
    };
}